web/requirements.dev.txt
django-debug-toolbar
django-livereload-server
ipython
debugpy
